refactor ui trace graph api
